eliminar el paquete fulllogic y llevar todo al paquete moredetails
