+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="ayds.winchester.songinfo.moredetails.fulllogic.view.MoreDetailsViewActivity"
+        <activity android:name="ayds.winchester.songinfo.moredetails.view.MoreDetailsViewActivity"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
