+++ app/src/main/java/ayds/winchester/songinfo/home/view/HomeView.kt
@@ -18,7 +18,7 @@ import ayds.winchester.songinfo.home.model.entities.EmptySong
 import ayds.winchester.songinfo.home.model.entities.Song
 import ayds.winchester.songinfo.home.model.entities.SpotifySong
 import ayds.winchester.songinfo.home.view.HomeUiState.Companion.DEFAULT_IMAGE
-import ayds.winchester.songinfo.moredetails.fulllogic.view.MoreDetailsViewActivity
+import ayds.winchester.songinfo.moredetails.view.MoreDetailsViewActivity
 import ayds.winchester.songinfo.utils.UtilsInjector
 import ayds.winchester.songinfo.utils.navigation.NavigationUtils
 import ayds.winchester.songinfo.utils.view.ImageLoader
