--- app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/controller/MoreDetailsController.kt
@@ -1,40 +1,0 @@
-package ayds.winchester.songinfo.moredetails.fulllogic.controller
-
-import ayds.observer.Observer
-import ayds.winchester.songinfo.moredetails.fulllogic.model.MoreDetailsModel
-import ayds.winchester.songinfo.moredetails.fulllogic.view.MoreDetailsUiEvent
-import ayds.winchester.songinfo.moredetails.fulllogic.view.MoreDetailsView
-
-interface MoreDetailsController {
-    fun setMoreDetailsView (moreDetailsView : MoreDetailsView)
-}
-
-internal class MoreDetailsControllerImpl(
-    private val moreDetailsModel : MoreDetailsModel
-) : MoreDetailsController {
-
-    private lateinit var moreDetailsView : MoreDetailsView
-
-    override fun setMoreDetailsView(moreDetailsView: MoreDetailsView) {
-        this.moreDetailsView = moreDetailsView
-        moreDetailsView.uiEventObservable.subscribe(observer)
-    }
-
-    private val observer: Observer<MoreDetailsUiEvent> =
-        Observer { value ->
-            when (value) {
-                MoreDetailsUiEvent.ViewFullArticle -> viewFullArticle()
-                MoreDetailsUiEvent.Search -> searchArtist()
-            }
-        }
-
-    private fun searchArtist() {
-        Thread {
-            moreDetailsModel.searchArtistInfo(moreDetailsView.uiState.artistName)
-        }.start()
-    }
-
-    private fun viewFullArticle() {
-        moreDetailsView.openFullArticle()
-    }
-}
