--- app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/view/MoreDetailsUiEvent.kt
@@ -1,6 +1,0 @@
-package ayds.winchester.songinfo.moredetails.fulllogic.view
-
-sealed class MoreDetailsUiEvent {
-    object ViewFullArticle : MoreDetailsUiEvent()
-    object Search : MoreDetailsUiEvent()
-}
