--- app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/view/MoreDetailsUiState.kt
@@ -1,10 +1,0 @@
-package ayds.winchester.songinfo.moredetails.fulllogic.view
-
-data class MoreDetailsUiState(
-    val artistName: String = "",
-    val info: String = "",
-    val pageid: String = "",
-    val actionsEnabled: Boolean = false,
-){
-
-}
