+++ app/src/main/java/ayds/winchester/songinfo/moredetails/model/repository/external/wikipedia/WikipediaArtistInfoService.kt
@@ -1,0 +1,8 @@
+package ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia
+
+import ayds.winchester.songinfo.moredetails.model.entities.WikipediaArtistInfo
+
+interface WikipediaArtistInfoService {
+
+    fun getArtistInfo(artistName: String?): WikipediaArtistInfo?
+}
