+++ app/src/main/java/ayds/winchester/songinfo/moredetails/model/repository/external/wikipedia/WikipediaInjector.kt
@@ -1,0 +1,8 @@
+package ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia
+
+import ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia.artistinfo.WikipediaArtistInfoInjector
+
+object WikipediaInjector {
+
+    val wikipediaArtistInfoService: WikipediaArtistInfoService = WikipediaArtistInfoInjector.wikipediaArtistInfoService
+}
