+++ app/src/main/java/ayds/winchester/songinfo/moredetails/view/MoreDetailsUiEvent.kt
@@ -1,0 +1,6 @@
+package ayds.winchester.songinfo.moredetails.view
+
+sealed class MoreDetailsUiEvent {
+    object ViewFullArticle : MoreDetailsUiEvent()
+    object Search : MoreDetailsUiEvent()
+}
