+++ app/src/main/java/ayds/winchester/songinfo/moredetails/view/MoreDetailsUiState.kt
@@ -1,0 +1,10 @@
+package ayds.winchester.songinfo.moredetails.view
+
+data class MoreDetailsUiState(
+    val artistName: String = "",
+    val info: String = "",
+    val pageid: String = "",
+    val actionsEnabled: Boolean = false,
+){
+
+}
